requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
Flask==1.1.2
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-language==2.0.0
googleapis-common-protos==1.52.0
grpcio==1.33.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
libcst==0.3.14
MarkupSafe==1.1.1
mypy-extensions==0.4.3
proto-plus==1.11.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-decouple==3.3
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
rsa==4.6
six==1.15.0
soupsieve==2.0.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
Werkzeug==1.0.1
